requirements.txt
Django>=5.0.7
django-filter
djangorestframework
django-cors-headers
djangorestframework_simplejwt
python-dotenv